Track animations
Added support for animated tracks.

Two methods can be used, the first is simply switching between AnimTrackModel1, AnimTrackModel2, and AnimTrackModel3. This is rather hacky and somewhat limited but this would be easier to implement via toolbox. The Leopard 2 model uses this as demonstration.

The second uses a proper array to store as many animation frames as necessary. The Abrams model uses this as demonstration.